exception handling for delete/searchById/update

+++ src/main/java/dev/Practice/DeliverySystem/model/services/UserService.java
@@ -1,12 +1,18 @@
 package dev.Practice.DeliverySystem.model.services;
 
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import dev.Practice.DeliverySystem.model.DTO.UserFullDTO;
 import dev.Practice.DeliverySystem.model.entities.User;
 import dev.Practice.DeliverySystem.model.repositories.UserRepository;
+import dev.Practice.DeliverySystem.model.services.servicesexceptions.DataBaseException;
+import dev.Practice.DeliverySystem.model.services.servicesexceptions.ResourceNotFoundException;
+
 import java.util.List;
 @Service
 public class UserService {
@@ -35,21 +41,37 @@ public class UserService {
     }
 
     @Transactional
-    public void deleteUser(Long id){ 
+    public void deleteUser(Long id){
+    if (repository.existsById(id)) {
         try{
-            repository.deleteById(id);
+        repository.deleteById(id);
+        }
+        catch(DataIntegrityViolationException e){
+            throw new DataBaseException("error! operation not enable");
+        }
+    }
+    else{
+        throw new ResourceNotFoundException(id);
+    }
+         
+  
+         
+        
        
-        }
-        catch(IllegalArgumentException e){
-          e.getStackTrace();
-          e.getMessage();
-        }
     }
     @Transactional
     public User UpdateUser(Long id, User obj){
-    User entity = repository.getReferenceById(id);
+        User entity = null;
+        try{
+     entity = repository.getReferenceById(id); //chamada para criar um objeto em persistencia 
     updateData(entity, obj);
     return repository.save(entity);
+    }
+
+    catch(EntityNotFoundException e){
+        throw new ResourceNotFoundException(id);
+    }
+    
     }
 
     private User updateData(User entity, User obj) {
